+++ tests/app/Opulence/Bootstrappers/Mocks/Bootstrapper.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2015 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Bootstrappers\Mocks;
+
+use Opulence\Bootstrappers\Bootstrapper as BaseBootstrapper;
+
+/**
+ * Mocks a bootstrapper for use in tests
+ */
+class Bootstrapper extends BaseBootstrapper
+{
+    // Don't do anything
+}