+++ tests/app/Opulence/Bootstrappers/Mocks/InvalidBootstrapper.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2015 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Bootstrappers\Mocks;
+
+/**
+ * Mocks an invalid bootstrapper for use in testing
+ */
+class InvalidBootstrapper
+{
+    // Don't do anything
+}