+++ tests/app/Opulence/Bootstrappers/Mocks/LazyFooInterface.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2015 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Bootstrappers\Mocks;
+
+/**
+ * Defines an interface for a class that's used by bootstrappers
+ */
+interface LazyFooInterface
+{
+    /**
+     * Gets the name of the concrete class
+     *
+     * @return string The name of the class
+     */
+    public function getClass();
+}